ENH: Adding to itk::Image the implementation of the method GetNumberOfComponentsPerPixel().      This method is provided in itkImageBase and overloaded in itkVectorImage, but was not      implemented in itkImage. With the current change, an itkImage of RGB pixels should      return 3 as the number of components per pixel.

+++ Testing/Code/Common/itkVectorImageTest.cxx
@@ -53,7 +53,7 @@ int itkVectorImageTest( int, char* argv[] )
 
   const unsigned int VectorLength = 6;
   const unsigned int Dimension    = 3;
-  typedef float PixelType;
+  typedef unsigned short PixelType;
 
   {
   // Test 1.
